# refactor: replace slice-based iteration with go 1.23 iterators